feat/apiSeparaion token 관련 모듈
