+++ src/api/pos/item/ItemGetApi.js
@@ -1,23 +1,10 @@
 import axios from 'axios';
 import URL from '../../BaseUrl';
+import { Token } from '../../Token';
 
 /* eslint-disable */
-
-// 로컬스토리지에 있는 토큰을 가져옴
-function getTokenFromLocalStorage(tokenName) {
-  return localStorage.getItem(tokenName);
-}
-
 export async function ItemGetApi(){
-
-    const authToken = getTokenFromLocalStorage('access_token');
-    const csrfToken = getTokenFromLocalStorage('csrf_token');
-
-    const headers = {
-        'content-type': 'application/json',
-        Authorization : `Bearer ${authToken}`,
-        'X-CSRF-Token' : csrfToken
-    };
+    const headers = Token()
 
     try{
         const res = await axios.get(`${URL}/categories/items`, {headers})
